# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API REST de videojuegos
  description: API REST para consultar y gestionar videojuegos, desarrolladores, reviews y países relacionados con sus desarrolladores.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Servidor local de desarrollo
...
paths:
...
  /games/{id}/reviews:
    get:
      tags:
        - Games
        - Reviews
      summary: Obtener reviews de un videojuego
      description: Devuelve las reviews asociadas al videojuego indicado.
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Reviews asociadas al videojuego.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameReviewsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: Obtener reviews
      description: Consulta las reviews guardadas en MongoDB, con filtros opcionales por ID y valoración.
      parameters:
        - name: id
          in: query
          description: ID numérico de la review.
          schema:
            type: integer
            minimum: 1
          example: 4
        - name: rating
          in: query
          description: Valoración de la review.
          schema:
            type: integer
            minimum: 1
            maximum: 5
          example: 5
        - name: page
          in: query
          description: Número de página.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 2
        - name: limit
          in: query
          description: Número máximo de resultados.
          schema:
            type: integer
            minimum: 1
            default: 1000
          example: 20
      responses:
        "200":
          description: Lista de reviews.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Reviews
      summary: Crear review
      description: Registra una nueva review asociada a un videojuego existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review creada correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedReviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Ya existe una review con ese ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /reviews/game/{gameId}:
    get:
      tags:
        - Reviews
      summary: Obtener reviews por videojuego
      description: Recupera las reviews vinculadas al videojuego indicado.
      parameters:
        - name: gameId
          in: path
          required: true
          description: ID numérico del videojuego.
          schema:
            type: integer
            minimum: 1
          example: 3328
      responses:
        "200":
          description: Reviews del videojuego.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsByGameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /reviews/{id}:
    patch:
      tags:
        - Reviews
      summary: Actualizar review
      description: Modifica parcialmente una review usando su ID propio.
      parameters:
        - name: id
          in: path
          required: true
          description: ID numérico de la review.
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Review actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedReviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Reviews
      summary: Eliminar review
      description: Borra la review identificada por su ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID numérico de la review.
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        "200":
          description: Review eliminada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    GameId:
      name: id
      in: path
      required: true
      description: ID numérico del videojuego.
      schema:
        type: integer
        minimum: 1
      example: 3328
  responses:
    BadRequest:
      description: Petición incorrecta.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: integer
          example: 1
        gameId:
          type: integer
          example: 3328
        gameName:
          type: string
          example: "The Witcher 3: Wild Hunt"
        user:
          type: string
          example: player_01
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Muy buen juego, lo recomiendo.
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required:
        - id
        - gameId
        - gameName
        - rating
        - comment
      properties:
        id:
          type: integer
          example: 4000
        gameId:
          type: integer
          example: 3328
        gameName:
          type: string
          example: "The Witcher 3: Wild Hunt"
        user:
          type: string
          example: player_01
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Muy buen juego, lo recomiendo.
    ReviewUpdate:
      type: object
      description: Campos admitidos en la actualización parcial de una review. No se actualizan id ni createdAt.
      properties:
        gameId:
          type: integer
        gameName:
          type: string
        user:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
...
    ReviewsListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 3482
        reviews_length:
          type: integer
          example: 20
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    ReviewsByGameResponse:
      type: object
      properties:
        reviews_length:
          type: integer
          example: 3
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    GameReviewsResponse:
      type: object
      properties:
        gameId:
          type: integer
          example: 3328
        gameName:
          type: string
          example: "The Witcher 3: Wild Hunt"
        reviews_length:
          type: integer
          example: 3
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
...
    CreatedReviewResponse:
      type: object
      properties:
        message:
          type: string
          example: Review creada correctamente
        id:
          type: integer
          example: 3
        review:
          $ref: "#/components/schemas/Review"
...
    UpdatedReviewResponse:
      type: object
      properties:
        message:
          type: string
          example: Review actualizada correctamente
        review:
          $ref: "#/components/schemas/Review"
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: El id debe ser un número positivo
        error:
          type: object
          description: Detalle técnico cuando el error procede de una excepción.
          additionalProperties: true
          nullable: true
        missingFields:
          type: array
          description: Campos obligatorios que faltan en la petición.
          items:
            type: string
        invalidFields:
          type: array
          description: Campos enviados que la ruta no admite.
          items:
            type: string